requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
langgraph==0.0.20
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
pypdf2==3.0.1
faiss-cpu==1.7.4
openai>=1.6.1,<2.0.0
